refatorando comands
