+++ src/Commands/Contracts/Generator.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace FaceDigital\FaceGen\Commands\Contracts;
+
+interface Generator
+{
+    public function handle(): void;
+    public function compileStub(): string;
+}
